services/massive/requirements.txt
redis==5.1.0
requests==2.32.5
websocket-client==1.8.0
polygon-api-client==1.16.3